update ai interview
